minor change based on code review feedback

+++ vtkm/filter/testing/UnitTestImageConnectivityFilter.cxx
@@ -68,7 +68,7 @@ void TestImageConnectivity()
   for (vtkm::Id i = 0; i < resultArrayHandle.GetNumberOfValues(); ++i)
   {
     VTKM_TEST_ASSERT(
-      test_equal(resultArrayHandle.GetPortalConstControl().Get(i), componentExpected[i]),
+      test_equal(resultArrayHandle.GetPortalConstControl().Get(i), componentExpected[size_t(i)]),
       "Wrong result for ImageConnectivity");
   }
 }
